[SIDE-1196] Add mobile navigation drop down
