+++ docs/src/components/menu/menu.tsx
@@ -1,0 +1,82 @@
+import React, { ChangeEvent } from "react"
+import {
+  buildAptitudeId,
+  buildAptitudePath,
+  buildCapabilityPath,
+} from "../aptitudes/aptitudePaths"
+import styles from "./menu.module.scss"
+import { Link, navigate } from "gatsby"
+import { IAptitudeData } from "../aptitudes/aptitudeData"
+
+interface IMenuProps {
+  aptitudes: IAptitudeData[]
+  currentPath: string
+}
+
+interface IMenuAptitudeProps {
+  aptitude: IAptitudeData
+  current: boolean
+}
+
+export const MenuAptitude: React.FunctionComponent<IMenuAptitudeProps> = props => {
+  const sensor = props.aptitude
+  const capabilities = sensor.capabilities.map(capability => {
+    return (
+      <li className={styles.sectionSubItem}>
+        <Link to={buildCapabilityPath(capability, sensor)}>
+          {capability.name}
+        </Link>
+      </li>
+    )
+  })
+
+  return (
+    <li className={styles.sectionItem}>
+      <Link to={buildAptitudePath(sensor)}>
+        <h2 className={styles.sectionItemHeader}>{sensor.name}</h2>
+      </Link>
+      {props.current && (
+        <ul className={styles.sectionSubItems}>{capabilities}</ul>
+      )}
+    </li>
+  )
+}
+
+export const MobileMenuSelect: React.FunctionComponent<IMenuProps> = props => {
+  const onChange = (newValue: ChangeEvent<HTMLSelectElement>) => {
+    navigate(newValue.target.value)
+  }
+  const sensorOptions = props.aptitudes.map(apt => (
+    <option value={buildAptitudePath(apt)}>{apt.name}</option>
+  ))
+  return (
+    <select onChange={onChange} value={props.currentPath}>
+      <optgroup label="Aptitudes">{sensorOptions}</optgroup>
+    </select>
+  )
+}
+
+export const Menu: React.FunctionComponent<IMenuProps> = props => {
+  let elements = props.aptitudes.map(aptitude => {
+    const aptitudeId = buildAptitudePath(aptitude)
+    return (
+      <MenuAptitude
+        aptitude={aptitude}
+        key={aptitudeId}
+        current={props.currentPath == aptitudeId}
+      />
+    )
+  })
+  return (
+    <div className={styles.menu}>
+      <MobileMenuSelect
+        aptitudes={props.aptitudes}
+        currentPath={props.currentPath}
+      />
+      <section className={styles.menuSection}>
+        <h1 className={styles.sectionTitle}>Aptitudes</h1>
+        <ul className={styles.sectionItems}>{elements}</ul>
+      </section>
+    </div>
+  )
+}
